api/requirements.txt
asgiref==3.6.0
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
coverage==6.5.0
DateTime==4.7
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
exceptiongroup==1.0.0
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.57.0
httplib2==0.21.0
idna==3.4
iniconfig==1.1.1
oauthlib==3.2.2
packaging==21.3
pluggy==1.0.0
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytest==7.2.0
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.12
zope.interface==5.5.0
